Refactor omnisci_backend to heavyai and move tests to heavyai folder
