+++ rbc/heavyai/__init__.py
@@ -7,9 +7,9 @@ from .omnisci_pipeline import *  # noqa: F401, F403
 from .omnisci_column_list import *  # noqa: F401, F403
 from .omnisci_table_function_manager import *  # noqa: F401, F403
 
-import rbc.omnisci_backend.mathimpl as math  # noqa: F401
-import rbc.omnisci_backend.npyimpl as np  # noqa: F401
-import rbc.omnisci_backend.python_operators as operators  # noqa: F401
+import rbc.heavyai.mathimpl as math  # noqa: F401
+import rbc.heavyai.npyimpl as np  # noqa: F401
+import rbc.heavyai.python_operators as operators  # noqa: F401
 
 # initialize the array api
 from rbc.stdlib import array_api  # noqa: F401
