+++ rbc/heavyai/npyimpl.py
