+++ rbc/heavyai/omnisci_bytes.py
