+++ rbc/heavyai/omnisci_table_function_manager.py
