+++ rbc/heavyai/omnisci_text_encoding.py
