+++ rbc/stdlib/statistical_functions.py
@@ -5,7 +5,7 @@ https://data-apis.org/array-api/latest/API_specification/statistical_functions.h
 """
 from rbc.externals.stdio import printf
 from rbc import typesystem
-from rbc.omnisci_backend import ArrayPointer
+from rbc.heavyai import ArrayPointer
 from rbc.stdlib import Expose
 from numba.core import extending, types, errors
 from numba.np import numpy_support
