+++ rbc/tests/heavyai/test_omnisci_array_null.py
