+++ rbc/tests/heavyai/test_omnisci_caller.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 from rbc.externals.omniscidb import set_output_row_size
-from rbc.omnisci_backend import Bytes
+from rbc.heavyai import Bytes
 from rbc.tests import omnisci_fixture, assert_equal
