+++ rbc/tests/heavyai/test_omnisci_column_null.py
