+++ rbc/tests/heavyai/test_omnisci_device_selection.py
