+++ rbc/tests/heavyai/test_omnisci_text_encoding.py
