+++ rbc/tests/stdlib/test_array_api.py
@@ -1,5 +1,5 @@
 from rbc.stdlib import array_api as xp
-from rbc.omnisci_backend.omnisci_buffer import free_buffer
+from rbc.heavyai.omnisci_buffer import free_buffer
 from ..test_rbclib import djit  # noqa: F401
